requirements.txt
coreapi==2.0.9
coverage==4.2
Django==1.10.2
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.1
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.1.7
Pygments==2.1.3
requests==2.11.1
simplejson==3.10.0
uritemplate==3.0.0
